refactor(web): remove cookbooks page
- Delete cookbooks page directory and API route
- Remove cookbooks from sitemap, middleware, and navigation
- Clean up related CSS styles and i18n translations
- Move Particles component to skills directory
