+++ turbo/apps/web/app/landing.css
@@ -1744,10 +1744,6 @@ section {
   border: 1px solid var(--border-light);
 }
 
-[data-theme="dark"] .cookbook-card {
-  border: none !important;
-}
-
 .feature-card.reverse {
   direction: rtl;
 }
